test-requirements.txt
typing-extensions
mock
coveralls
django-storages
boto3
google-cloud-storage
pytest
pytest-mock
pytest-django
django-stubs
boto3-stubs[s3]
types-s3transfer
